+++ extensions/material-helper/src/utils/getComponentSource.ts
@@ -1,0 +1,32 @@
+import * as parser from '@babel/parser';
+import traverse from '@babel/traverse';
+import getBabelParserPlugins from './getBabelParserPlugins';
+
+function getComponentSource(
+  documentText: string,
+  tagName: string,
+) {
+  const ast = parser.parse(documentText, {
+    sourceType: 'module',
+    plugins: getBabelParserPlugins('jsx'),
+  });
+
+  const result = { source: '', importedComponent: '' };
+
+  traverse(ast, {
+    ImportDeclaration(path) {
+      const specifiers = path.get('specifiers');
+      const targetSpecifier = specifiers.find(specifier => specifier.node.local.name === tagName);
+      if (targetSpecifier) {
+        result.source = path.node.source.value;
+        // @ts-ignore
+        result.importedComponent = targetSpecifier.node.imported.name;
+        path.stop();
+      }
+    }
+  })
+
+  return result;
+}
+
+export default getComponentSource;