+++ packages/doctor/src/config.ts
@@ -5,6 +5,7 @@ export default {
       eslint: 'eslint-result.json',
       escomplex: 'escomplex-result.json',
       jscpd: 'jscpd-result.json',
+      codemod: 'codemod-result.json',
     },
   },
 };