+++ extensions/material-helper/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
+
+## 1.0.6
+
+- fix: command `Find Components In Current File` not working.
+- fix: hover component link
+
 ## 1.0.3 ~ 1.0.5
 
 - fix: rebuild and fix .vscodeignore configuration [#576](https://github.com/microsoft/vscode-vsce/issues/576)